dags/requirements.txt
attrs==19.3.0
Automat==0.8.0
Babel==2.10.3
backcall==0.2.0
backports.entry-points-selectable==1.1.1
blinker==1.5
cachelib==0.9.0
cattrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
clickclick==20.10.2
cloud-init==22.1
colorama==0.4.5
colorlog==4.8.0
command-not-found==0.3
commonmark==0.9.1
configobj==5.0.6
ConfigUpdater==3.1.1
connexion==2.14.1
constantly==15.1.0
cron-descriptor==1.2.31
croniter==1.3.7
cryptography==36.0.2
cupshelpers==1.0
dbus-python==1.2.16
decorator==5.1.1
defer==1.0.6
Deprecated==1.2.13
dill==0.3.1.1
distlib==0.3.3
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.2.1
docutils==0.19
email-validator==1.3.0
entrypoints==0.3
exceptiongroup==1.0.0rc9
executing==1.0.0
filelock==3.4.0
Flask==2.2.2
Flask-AppBuilder==4.1.3
Flask-Babel==2.0.0
Flask-Caching==2.0.1
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
graphviz==0.20.1
greenlet==1.1.3
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httplib2==0.14.0
httpx==0.23.0
hyperlink==19.0.0
idna==3.4
importlib-metadata==4.12.0
importlib-resources==5.9.0
incremental==16.10.1
inflection==0.5.1
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==4.16.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.7.1
linecache2==1.0.0
linkify-it-py==2.0.0
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.2.3
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.1
mdurl==0.1.2
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.23.2
oauthlib==3.1.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pathspec==0.9.0
pbr==5.4.5
pendulum==2.1.2
pexpect==4.6.0
pickleshare==0.7.5
pkgutil_resolve_name==1.3.10
platformdirs==2.4.0
pluggy==1.0.0
ply==3.11
prison==0.2.1
prompt-toolkit==3.0.31
protobuf==3.6.1
psutil==5.9.2
pure-eval==0.2.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
Pygments==2.13.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==2.5.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.7
python-daemon==2.3.1
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.21.0
python-nvd3==0.15.0
python-slugify==6.1.2
pytz==2022.2.1
pytzdata==2020.1
PyYAML==6.0
requests==2.28.1
requests-cache==0.9.6
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rfc3986==1.5.0
rich==12.5.1
SecretStorage==2.3.1
service-identity==18.1.0
setproctitle==1.3.2
simplejson==3.16.0
six==1.16.0
sniffio==1.3.0
sos==4.3
SQLAlchemy==1.4.27
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.3
sqlparse==0.4.3
ssh-import-id==5.10
stack-data==0.5.0
swagger-ui-bundle==0.0.9
systemd-python==234
tabulate==0.8.10
tenacity==8.1.0
termcolor==2.0.1
testresources==2.0.0
text-unidecode==1.3
traceback2==1.4.0
traitlets==5.3.0
Twisted==18.9.0
typing_extensions==4.3.0
ubuntu-advantage-tools==27.8
uc-micro-py==1.0.1
ufw==0.36
unattended-upgrades==0.1
unicodecsv==0.14.1
unittest2==1.1.0
url-normalize==1.4.3
urllib3==1.26.12
virtualenv==20.10.0
wadllib==1.3.3
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==2.3.3
zipp==3.8.1
zope.interface==4.7.1
